requirements.txt
fastapi==0.70.1
matplotlib==3.5.1
Pillow==8.4.0
uvicorn==0.16.0
